@@ -5,21 +5,23 @@ publish_to: none
 
 environment:
   sdk: ">=2.16.0 <3.0.0"
-  flutter: ">=2.10.2"
+  flutter: ">=3.0.0"
 
 dependencies:
+  api_models:
+    git:
+      url: https://github.com/btrautmann/irri-models.git
+      ref: main
   bloc: ^8.0.3
   clock: ^1.1.0
-  cloud_firestore: ^3.1.0
   collection: ^1.15.0
   core:
     path: ../core
   core_ui:
     path: ../core_ui
-  dio: ^4.0.0
+  dio: ^4.0.6
   equatable: ^2.0.3
   firebase_analytics: ^9.1.2
-  firebase_auth: ^3.3.9
   firebase_core: ^1.13.1
   firebase_crashlytics: ^2.2.4
   firebase_messaging: ^11.2.8
@@ -29,12 +31,10 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_native_timezone: ^2.0.0
-  freezed_annotation: ^1.1.0
+  freezed_annotation: ^2.0.3
   go_router: ^3.0.4
   google_fonts: ^2.1.0
   http: ^0.13.3
-  hydrawise:
-    path: ../hydrawise
   intl: ^0.17.0
   path: ^1.8.0
   pedantic: ^1.11.1
@@ -54,7 +54,7 @@ dev_dependencies:
   flutter_lints: ^1.0.4
   flutter_test:
     sdk: flutter
-  freezed: ^1.1.0
+  freezed: ^2.0.5
   json_serializable: ^6.1.4
   mocktail: ^0.3.0
 
