@@ -5,7 +5,7 @@ publish_to: none
 
 environment:
   sdk: ">=2.16.2 <3.0.0"
-  flutter: ">=1.17.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   bloc: ^8.0.3
@@ -16,7 +16,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_bloc: ^8.0.1
-  freezed_annotation: ^1.1.0
+  freezed_annotation: ^2.0.3
   result_type: ^0.1.0
   weather: ^2.0.1
 
@@ -24,6 +24,6 @@ dev_dependencies:
   flutter_lints: ^1.0.0
   flutter_test:
     sdk: flutter
-  freezed: ^1.1.1
+  freezed: ^2.0.5
 
 flutter:
