@@ -1,14 +1,14 @@
+import 'package:api_models/api_models.dart';
 import 'package:charlatan/charlatan.dart';
 import 'package:core/core.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:hydrawise/hydrawise.dart';
 import 'package:irri/auth/auth.dart';
 
 import '../../core/fakes/fake_http_client.dart';
 
 void main() {
-  late ValidateApiKey subject;
+  late LogIn subject;
   final storage = InMemoryStorage();
   final setApiKey = SetApiKey(storage);
   final getApiKey = GetApiKey(storage);
@@ -16,7 +16,7 @@ void main() {
   group('call(:apiKey)', () {
     Future<void> _buildSubject(Charlatan charlatan) async {
       final client = FakeHttpClient(charlatan);
-      subject = ValidateApiKey(
+      subject = LogIn(
         httpClient: client,
         setApiKey: setApiKey,
       );
@@ -27,36 +27,38 @@ void main() {
     });
 
     test('it hits the api with correct parameters', () async {
-      late Map<String, Object?> queryParameters;
-      late RequestOptions options;
+      late final LoginRequest loginRequest;
+      late final RequestOptions options;
 
       final charlatan = Charlatan()
-        ..whenGet('customerdetails.php', (request) {
-          queryParameters = request.queryParameters;
+        ..whenPost('login', (request) {
+          loginRequest =
+              // ignore: cast_nullable_to_non_nullable
+              LoginRequest.fromJson(request.body as Map<String, dynamic>);
           options = request.requestOptions;
-          return CustomerDetails(
+          return Customer(
             activeControllerId: 1,
             customerId: 1,
-            controllers: [],
+            apiKey: 'fake-api-key',
           );
         });
       await _buildSubject(charlatan);
 
       await subject.call('fake-api-key');
 
-      expect(queryParameters['api_key'], 'fake-api-key');
-      expect(queryParameters['type'], 'controllers');
+      expect(loginRequest.apiKey, 'fake-api-key');
+      expect(loginRequest.type, 'controllers');
       expect(options.extra['allow_auth_errors'], true);
     });
 
     group('when api call succeeds', () {
       setUp(() async {
         final charlatan = Charlatan()
-          ..whenGet('customerdetails.php', (request) {
-            return CustomerDetails(
+          ..whenPost('login', (request) {
+            return Customer(
               activeControllerId: 1,
               customerId: 1,
-              controllers: [],
+              apiKey: 'fake-api-key',
             );
           });
         await _buildSubject(charlatan);
@@ -82,8 +84,8 @@ void main() {
     group('when api call fails', () {
       setUp(() async {
         final charlatan = Charlatan()
-          ..whenGet(
-            'customerdetails.php',
+          ..whenPost(
+            'login',
             (request) => CharlatanHttpResponse(statusCode: 500),
           );
         await _buildSubject(charlatan);
