@@ -1,6 +1,5 @@
-import 'package:clock/clock.dart';
+import 'package:api_models/api_models.dart';
 import 'package:dio/dio.dart';
-import 'package:hydrawise/hydrawise.dart';
 
 /// An [Interceptor] that intercepts requests and checks for the
 /// existence of a `demo` API key. If that key is present, we're
@@ -11,44 +10,42 @@ class DemoModeInterceptor extends Interceptor {
   void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
     final apiKey = options.queryParameters['api_key'];
     if (apiKey != null && apiKey == 'demo') {
-      if (options.uri.pathSegments.contains('customerdetails.php')) {
+      if (options.uri.pathSegments.contains('login')) {
         handler.resolve(
           Response(
             statusCode: 200,
             requestOptions: options,
-            data: CustomerDetails(
-              activeControllerId: 1,
-              customerId: 1,
-              controllers: [
-                Controller(
-                  id: 1,
-                  name: 'Fake Controller',
-                  lastContact: clock.now().millisecondsSinceEpoch,
-                  serialNumber: 'fake-serial-number',
-                  status: 'All good!',
-                )
-              ],
+            data: LoginResponse(
+              customer: Customer(
+                activeControllerId: 1,
+                customerId: 1,
+                apiKey: 'fake-api-key',
+              ),
             ).toJson(),
           ),
         );
-      } else if (options.uri.pathSegments.contains('statusschedule.php')) {
+      } else if (options.uri.pathSegments.contains('customer')) {
         handler.resolve(
           Response(
             statusCode: 200,
             requestOptions: options,
-            data: CustomerStatus(
-              numberOfSecondsUntilNextRequest: 60,
-              timeOfLastStatusUnixEpoch: 1635901976,
-              zones: [
-                Zone(
-                  id: 1,
-                  physicalNumber: 1,
-                  name: 'Fake Zone',
-                  nextTimeOfWaterFriendly: '7:00',
-                  secondsUntilNextRun: 10000,
-                  lengthOfNextRunTimeOrTimeRemaining: 900,
-                ),
-              ],
+            data: GetCustomerResponse(
+              customer: Customer(
+                activeControllerId: 1,
+                customerId: 1,
+                apiKey: 'fake-api-key',
+              ),
+              zones: List.of(
+                [
+                  Zone(
+                    id: 12345,
+                    number: 1,
+                    name: 'Fake Zone',
+                    timeUntilNextRunSec: 60,
+                    runLengthSec: 60,
+                  )
+                ],
+              ),
             ).toJson(),
           ),
         );
