@@ -1,5 +1,6 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'configuration_cubit.dart';
@@ -11,28 +12,7 @@ part of 'configuration_cubit.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-/// @nodoc
-class _$ConfigurationStateTearOff {
-  const _$ConfigurationStateTearOff();
-
-  _ConfigurationState call(
-      {bool pushNotificationsEnabled = false,
-      String? localTimezone = null,
-      List<String> availableTimezones = const [],
-      String? timeZone}) {
-    return _ConfigurationState(
-      pushNotificationsEnabled: pushNotificationsEnabled,
-      localTimezone: localTimezone,
-      availableTimezones: availableTimezones,
-      timeZone: timeZone,
-    );
-  }
-}
-
-/// @nodoc
-const $ConfigurationState = _$ConfigurationStateTearOff();
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 /// @nodoc
 mixin _$ConfigurationState {
@@ -96,11 +76,11 @@ class _$ConfigurationStateCopyWithImpl<$Res>
 }
 
 /// @nodoc
-abstract class _$ConfigurationStateCopyWith<$Res>
+abstract class _$$_ConfigurationStateCopyWith<$Res>
     implements $ConfigurationStateCopyWith<$Res> {
-  factory _$ConfigurationStateCopyWith(
-          _ConfigurationState value, $Res Function(_ConfigurationState) then) =
-      __$ConfigurationStateCopyWithImpl<$Res>;
+  factory _$$_ConfigurationStateCopyWith(_$_ConfigurationState value,
+          $Res Function(_$_ConfigurationState) then) =
+      __$$_ConfigurationStateCopyWithImpl<$Res>;
   @override
   $Res call(
       {bool pushNotificationsEnabled,
@@ -110,15 +90,15 @@ abstract class _$ConfigurationStateCopyWith<$Res>
 }
 
 /// @nodoc
-class __$ConfigurationStateCopyWithImpl<$Res>
+class __$$_ConfigurationStateCopyWithImpl<$Res>
     extends _$ConfigurationStateCopyWithImpl<$Res>
-    implements _$ConfigurationStateCopyWith<$Res> {
-  __$ConfigurationStateCopyWithImpl(
-      _ConfigurationState _value, $Res Function(_ConfigurationState) _then)
-      : super(_value, (v) => _then(v as _ConfigurationState));
+    implements _$$_ConfigurationStateCopyWith<$Res> {
+  __$$_ConfigurationStateCopyWithImpl(
+      _$_ConfigurationState _value, $Res Function(_$_ConfigurationState) _then)
+      : super(_value, (v) => _then(v as _$_ConfigurationState));
 
   @override
-  _ConfigurationState get _value => super._value as _ConfigurationState;
+  _$_ConfigurationState get _value => super._value as _$_ConfigurationState;
 
   @override
   $Res call({
@@ -127,7 +107,7 @@ class __$ConfigurationStateCopyWithImpl<$Res>
     Object? availableTimezones = freezed,
     Object? timeZone = freezed,
   }) {
-    return _then(_ConfigurationState(
+    return _then(_$_ConfigurationState(
       pushNotificationsEnabled: pushNotificationsEnabled == freezed
           ? _value.pushNotificationsEnabled
           : pushNotificationsEnabled // ignore: cast_nullable_to_non_nullable
@@ -137,7 +117,7 @@ class __$ConfigurationStateCopyWithImpl<$Res>
           : localTimezone // ignore: cast_nullable_to_non_nullable
               as String?,
       availableTimezones: availableTimezones == freezed
-          ? _value.availableTimezones
+          ? _value._availableTimezones
           : availableTimezones // ignore: cast_nullable_to_non_nullable
               as List<String>,
       timeZone: timeZone == freezed
@@ -154,18 +134,24 @@ class _$_ConfigurationState implements _ConfigurationState {
   _$_ConfigurationState(
       {this.pushNotificationsEnabled = false,
       this.localTimezone = null,
-      this.availableTimezones = const [],
-      this.timeZone});
+      final List<String> availableTimezones = const [],
+      this.timeZone})
+      : _availableTimezones = availableTimezones;
 
-  @JsonKey()
   @override
+  @JsonKey()
   final bool pushNotificationsEnabled;
-  @JsonKey()
   @override
+  @JsonKey()
   final String? localTimezone;
-  @JsonKey()
+  final List<String> _availableTimezones;
   @override
-  final List<String> availableTimezones;
+  @JsonKey()
+  List<String> get availableTimezones {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_availableTimezones);
+  }
+
   @override
   final String? timeZone;
 
@@ -178,13 +164,13 @@ class _$_ConfigurationState implements _ConfigurationState {
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _ConfigurationState &&
+            other is _$_ConfigurationState &&
             const DeepCollectionEquality().equals(
                 other.pushNotificationsEnabled, pushNotificationsEnabled) &&
             const DeepCollectionEquality()
                 .equals(other.localTimezone, localTimezone) &&
             const DeepCollectionEquality()
-                .equals(other.availableTimezones, availableTimezones) &&
+                .equals(other._availableTimezones, _availableTimezones) &&
             const DeepCollectionEquality().equals(other.timeZone, timeZone));
   }
 
@@ -193,21 +179,22 @@ class _$_ConfigurationState implements _ConfigurationState {
       runtimeType,
       const DeepCollectionEquality().hash(pushNotificationsEnabled),
       const DeepCollectionEquality().hash(localTimezone),
-      const DeepCollectionEquality().hash(availableTimezones),
+      const DeepCollectionEquality().hash(_availableTimezones),
       const DeepCollectionEquality().hash(timeZone));
 
   @JsonKey(ignore: true)
   @override
-  _$ConfigurationStateCopyWith<_ConfigurationState> get copyWith =>
-      __$ConfigurationStateCopyWithImpl<_ConfigurationState>(this, _$identity);
+  _$$_ConfigurationStateCopyWith<_$_ConfigurationState> get copyWith =>
+      __$$_ConfigurationStateCopyWithImpl<_$_ConfigurationState>(
+          this, _$identity);
 }
 
 abstract class _ConfigurationState implements ConfigurationState {
   factory _ConfigurationState(
-      {bool pushNotificationsEnabled,
-      String? localTimezone,
-      List<String> availableTimezones,
-      String? timeZone}) = _$_ConfigurationState;
+      {final bool pushNotificationsEnabled,
+      final String? localTimezone,
+      final List<String> availableTimezones,
+      final String? timeZone}) = _$_ConfigurationState;
 
   @override
   bool get pushNotificationsEnabled;
@@ -219,6 +206,6 @@ abstract class _ConfigurationState implements ConfigurationState {
   String? get timeZone;
   @override
   @JsonKey(ignore: true)
-  _$ConfigurationStateCopyWith<_ConfigurationState> get copyWith =>
+  _$$_ConfigurationStateCopyWith<_$_ConfigurationState> get copyWith =>
       throw _privateConstructorUsedError;
 }
