@@ -1,8 +1,9 @@
 import 'dart:async';
 
+import 'package:api_models/api_models.dart';
 import 'package:bloc/bloc.dart';
+import 'package:clock/clock.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:hydrawise/hydrawise.dart';
 import 'package:irri/auth/auth.dart';
 import 'package:irri/customer_details/customer_details.dart';
 
@@ -12,25 +13,28 @@ part 'customer_details_state.dart';
 class CustomerDetailsCubit extends Cubit<CustomerDetailsState> {
   CustomerDetailsCubit({
     required GetCustomerDetails getCustomerDetails,
-    required GetCustomerStatus getCustomerStatus,
+    required SetNextPollTime setNextPollTime,
+    required GetNextPollTime getNextPollTime,
     required AuthCubit authCubit,
   })  : _getCustomerDetails = getCustomerDetails,
-        _getCustomerStatus = getCustomerStatus,
+        _setNextPollTime = setNextPollTime,
+        _getNextPollTime = getNextPollTime,
         _authCubit = authCubit,
         super(CustomerDetailsState.loading());
 
   final GetCustomerDetails _getCustomerDetails;
-  final GetCustomerStatus _getCustomerStatus;
+  final SetNextPollTime _setNextPollTime;
+  final GetNextPollTime _getNextPollTime;
   final AuthCubit _authCubit;
 
   Timer? _timer;
 
-  void start() {
+  Future<void> start() async {
     if (_authCubit.state.isLoggedIn()) {
       // Need to check initial login state because
       // LoginCubit stream will not emit logged in
       // if already logged in when we begin listening
-      _loadCustomerDetails();
+      await _loadCustomerDetails();
     }
     _authCubit.stream.asBroadcastStream().listen((event) {
       event.when(
@@ -62,17 +66,12 @@ class CustomerDetailsCubit extends Cubit<CustomerDetailsState> {
 
     // TODO(brandon): Handle failure
     if (customerDetails.isSuccess) {
-      final customerStatus = await _getCustomerStatus(
-        activeControllerId: customerDetails.success.activeControllerId,
+      emit(
+        CustomerDetailsState.complete(
+          customerDetails: customerDetails.success.customer,
+          zones: customerDetails.success.zones,
+        ),
       );
-      if (customerStatus.isSuccess) {
-        emit(
-          CustomerDetailsState.complete(
-            customerDetails: customerDetails.success,
-            customerStatus: customerStatus.success,
-          ),
-        );
-      }
       _poll();
     }
   }
@@ -80,21 +79,26 @@ class CustomerDetailsCubit extends Cubit<CustomerDetailsState> {
   void _poll() {
     _timer = Timer.periodic(const Duration(seconds: 5), (timer) {
       state.maybeWhen(
-        complete: (details, status) async {
-          final customerStatus = await _getCustomerStatus();
-          if (customerStatus.isSuccess) {
-            final status = customerStatus.success;
-            status.zones.sort(
-              (z1, z2) => z1.physicalNumber.compareTo(
-                z2.physicalNumber,
-              ),
-            );
-            emit(
-              CustomerDetailsState.complete(
-                customerDetails: details,
-                customerStatus: status,
-              ),
-            );
+        complete: (details, zones) async {
+          final nextPollTime = await _getNextPollTime();
+          if (clock.now().isAfter(nextPollTime)) {
+            final response = await _getCustomerDetails();
+            if (response.isSuccess) {
+              // TODO(brandon): Build this back into the API
+              await _setNextPollTime(secondsUntilNextPoll: 60);
+              final getCustomerResponse = response.success;
+              List.of(getCustomerResponse.zones).sort(
+                (z1, z2) => z1.number.compareTo(
+                  z2.number,
+                ),
+              );
+              emit(
+                CustomerDetailsState.complete(
+                  customerDetails: getCustomerResponse.customer,
+                  zones: getCustomerResponse.zones,
+                ),
+              );
+            }
           }
         },
         orElse: () {},
