@@ -1,5 +1,6 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'customer_details_cubit.dart';
@@ -11,53 +12,27 @@ part of 'customer_details_cubit.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-/// @nodoc
-class _$CustomerDetailsStateTearOff {
-  const _$CustomerDetailsStateTearOff();
-
-  _Loading loading() {
-    return _Loading();
-  }
-
-  _Complete complete(
-      {required CustomerDetails customerDetails,
-      required CustomerStatus customerStatus}) {
-    return _Complete(
-      customerDetails: customerDetails,
-      customerStatus: customerStatus,
-    );
-  }
-}
-
-/// @nodoc
-const $CustomerDetailsState = _$CustomerDetailsStateTearOff();
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 /// @nodoc
 mixin _$CustomerDetailsState {
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() loading,
-    required TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)
+    required TResult Function(Customer customerDetails, List<Zone> zones)
         complete,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult Function()? loading,
-    TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)?
-        complete,
+    TResult Function(Customer customerDetails, List<Zone> zones)? complete,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? loading,
-    TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)?
-        complete,
+    TResult Function(Customer customerDetails, List<Zone> zones)? complete,
     required TResult orElse(),
   }) =>
       throw _privateConstructorUsedError;
@@ -100,20 +75,21 @@ class _$CustomerDetailsStateCopyWithImpl<$Res>
 }
 
 /// @nodoc
-abstract class _$LoadingCopyWith<$Res> {
-  factory _$LoadingCopyWith(_Loading value, $Res Function(_Loading) then) =
-      __$LoadingCopyWithImpl<$Res>;
+abstract class _$$_LoadingCopyWith<$Res> {
+  factory _$$_LoadingCopyWith(
+          _$_Loading value, $Res Function(_$_Loading) then) =
+      __$$_LoadingCopyWithImpl<$Res>;
 }
 
 /// @nodoc
-class __$LoadingCopyWithImpl<$Res>
+class __$$_LoadingCopyWithImpl<$Res>
     extends _$CustomerDetailsStateCopyWithImpl<$Res>
-    implements _$LoadingCopyWith<$Res> {
-  __$LoadingCopyWithImpl(_Loading _value, $Res Function(_Loading) _then)
-      : super(_value, (v) => _then(v as _Loading));
+    implements _$$_LoadingCopyWith<$Res> {
+  __$$_LoadingCopyWithImpl(_$_Loading _value, $Res Function(_$_Loading) _then)
+      : super(_value, (v) => _then(v as _$_Loading));
 
   @override
-  _Loading get _value => super._value as _Loading;
+  _$_Loading get _value => super._value as _$_Loading;
 }
 
 /// @nodoc
@@ -129,7 +105,7 @@ class _$_Loading implements _Loading {
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _Loading);
+        (other.runtimeType == runtimeType && other is _$_Loading);
   }
 
   @override
@@ -139,8 +115,7 @@ class _$_Loading implements _Loading {
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() loading,
-    required TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)
+    required TResult Function(Customer customerDetails, List<Zone> zones)
         complete,
   }) {
     return loading();
@@ -150,9 +125,7 @@ class _$_Loading implements _Loading {
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult Function()? loading,
-    TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)?
-        complete,
+    TResult Function(Customer customerDetails, List<Zone> zones)? complete,
   }) {
     return loading?.call();
   }
@@ -161,9 +134,7 @@ class _$_Loading implements _Loading {
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? loading,
-    TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)?
-        complete,
+    TResult Function(Customer customerDetails, List<Zone> zones)? complete,
     required TResult orElse(),
   }) {
     if (loading != null) {
@@ -209,53 +180,47 @@ abstract class _Loading implements CustomerDetailsState {
 }
 
 /// @nodoc
-abstract class _$CompleteCopyWith<$Res> {
-  factory _$CompleteCopyWith(_Complete value, $Res Function(_Complete) then) =
-      __$CompleteCopyWithImpl<$Res>;
-  $Res call({CustomerDetails customerDetails, CustomerStatus customerStatus});
+abstract class _$$_CompleteCopyWith<$Res> {
+  factory _$$_CompleteCopyWith(
+          _$_Complete value, $Res Function(_$_Complete) then) =
+      __$$_CompleteCopyWithImpl<$Res>;
+  $Res call({Customer customerDetails, List<Zone> zones});
 
-  $CustomerDetailsCopyWith<$Res> get customerDetails;
-  $CustomerStatusCopyWith<$Res> get customerStatus;
+  $CustomerCopyWith<$Res> get customerDetails;
 }
 
 /// @nodoc
-class __$CompleteCopyWithImpl<$Res>
+class __$$_CompleteCopyWithImpl<$Res>
     extends _$CustomerDetailsStateCopyWithImpl<$Res>
-    implements _$CompleteCopyWith<$Res> {
-  __$CompleteCopyWithImpl(_Complete _value, $Res Function(_Complete) _then)
-      : super(_value, (v) => _then(v as _Complete));
+    implements _$$_CompleteCopyWith<$Res> {
+  __$$_CompleteCopyWithImpl(
+      _$_Complete _value, $Res Function(_$_Complete) _then)
+      : super(_value, (v) => _then(v as _$_Complete));
 
   @override
-  _Complete get _value => super._value as _Complete;
+  _$_Complete get _value => super._value as _$_Complete;
 
   @override
   $Res call({
     Object? customerDetails = freezed,
-    Object? customerStatus = freezed,
+    Object? zones = freezed,
   }) {
-    return _then(_Complete(
+    return _then(_$_Complete(
       customerDetails: customerDetails == freezed
           ? _value.customerDetails
           : customerDetails // ignore: cast_nullable_to_non_nullable
-              as CustomerDetails,
-      customerStatus: customerStatus == freezed
-          ? _value.customerStatus
-          : customerStatus // ignore: cast_nullable_to_non_nullable
-              as CustomerStatus,
+              as Customer,
+      zones: zones == freezed
+          ? _value._zones
+          : zones // ignore: cast_nullable_to_non_nullable
+              as List<Zone>,
     ));
   }
 
   @override
-  $CustomerDetailsCopyWith<$Res> get customerDetails {
-    return $CustomerDetailsCopyWith<$Res>(_value.customerDetails, (value) {
+  $CustomerCopyWith<$Res> get customerDetails {
+    return $CustomerCopyWith<$Res>(_value.customerDetails, (value) {
       return _then(_value.copyWith(customerDetails: value));
-    });
-  }
-
-  @override
-  $CustomerStatusCopyWith<$Res> get customerStatus {
-    return $CustomerStatusCopyWith<$Res>(_value.customerStatus, (value) {
-      return _then(_value.copyWith(customerStatus: value));
     });
   }
 }
@@ -263,73 +228,72 @@ class __$CompleteCopyWithImpl<$Res>
 /// @nodoc
 
 class _$_Complete implements _Complete {
-  _$_Complete({required this.customerDetails, required this.customerStatus});
+  _$_Complete({required this.customerDetails, required final List<Zone> zones})
+      : _zones = zones;
 
   @override
-  final CustomerDetails customerDetails;
+  final Customer customerDetails;
+  final List<Zone> _zones;
   @override
-  final CustomerStatus customerStatus;
+  List<Zone> get zones {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_zones);
+  }
 
   @override
   String toString() {
-    return 'CustomerDetailsState.complete(customerDetails: $customerDetails, customerStatus: $customerStatus)';
+    return 'CustomerDetailsState.complete(customerDetails: $customerDetails, zones: $zones)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _Complete &&
+            other is _$_Complete &&
             const DeepCollectionEquality()
                 .equals(other.customerDetails, customerDetails) &&
-            const DeepCollectionEquality()
-                .equals(other.customerStatus, customerStatus));
+            const DeepCollectionEquality().equals(other._zones, _zones));
   }
 
   @override
   int get hashCode => Object.hash(
       runtimeType,
       const DeepCollectionEquality().hash(customerDetails),
-      const DeepCollectionEquality().hash(customerStatus));
+      const DeepCollectionEquality().hash(_zones));
 
   @JsonKey(ignore: true)
   @override
-  _$CompleteCopyWith<_Complete> get copyWith =>
-      __$CompleteCopyWithImpl<_Complete>(this, _$identity);
+  _$$_CompleteCopyWith<_$_Complete> get copyWith =>
+      __$$_CompleteCopyWithImpl<_$_Complete>(this, _$identity);
 
   @override
   @optionalTypeArgs
   TResult when<TResult extends Object?>({
     required TResult Function() loading,
-    required TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)
+    required TResult Function(Customer customerDetails, List<Zone> zones)
         complete,
   }) {
-    return complete(customerDetails, customerStatus);
+    return complete(customerDetails, zones);
   }
 
   @override
   @optionalTypeArgs
   TResult? whenOrNull<TResult extends Object?>({
     TResult Function()? loading,
-    TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)?
-        complete,
+    TResult Function(Customer customerDetails, List<Zone> zones)? complete,
   }) {
-    return complete?.call(customerDetails, customerStatus);
+    return complete?.call(customerDetails, zones);
   }
 
   @override
   @optionalTypeArgs
   TResult maybeWhen<TResult extends Object?>({
     TResult Function()? loading,
-    TResult Function(
-            CustomerDetails customerDetails, CustomerStatus customerStatus)?
-        complete,
+    TResult Function(Customer customerDetails, List<Zone> zones)? complete,
     required TResult orElse(),
   }) {
     if (complete != null) {
-      return complete(customerDetails, customerStatus);
+      return complete(customerDetails, zones);
     }
     return orElse();
   }
@@ -368,12 +332,12 @@ class _$_Complete implements _Complete {
 
 abstract class _Complete implements CustomerDetailsState {
   factory _Complete(
-      {required CustomerDetails customerDetails,
-      required CustomerStatus customerStatus}) = _$_Complete;
+      {required final Customer customerDetails,
+      required final List<Zone> zones}) = _$_Complete;
 
-  CustomerDetails get customerDetails;
-  CustomerStatus get customerStatus;
+  Customer get customerDetails;
+  List<Zone> get zones;
   @JsonKey(ignore: true)
-  _$CompleteCopyWith<_Complete> get copyWith =>
+  _$$_CompleteCopyWith<_$_Complete> get copyWith =>
       throw _privateConstructorUsedError;
 }
