@@ -5,7 +5,7 @@ class CustomerDetailsState with _$CustomerDetailsState {
   factory CustomerDetailsState.loading() = _Loading;
 
   factory CustomerDetailsState.complete({
-    required CustomerDetails customerDetails,
-    required CustomerStatus customerStatus,
+    required Customer customerDetails,
+    required List<Zone> zones,
   }) = _Complete;
 }
