@@ -1,49 +1,33 @@
-import 'package:hydrawise/hydrawise.dart';
-import 'package:irri/programs/programs.dart';
+import 'package:api_models/api_models.dart';
 
 abstract class CustomerDetailsRepository {
+  // Customer
   Future<void> insertCustomer(Customer customer);
-  Future<void> updateCustomer(CustomerStatus customerStatus);
-  Future<void> updateCustomerTimeZone(String timeZone);
   Future<Customer?> getCustomer();
+  Future<void> updateCustomerTimeZone(String timeZone);
+
+  // FCM Tokens
   Future<void> addFcmToken(String token);
   Future<void> removeFcmToken(String token);
   Future<List<String>> getRegisteredFcmTokens();
+
+  // Timezone
   Future<void> updateTimeZone(String timeZone);
   Future<String?> getUserTimeZone();
+
+  // Zones
   Future<void> insertZone(Zone zone);
   Future<void> updateZone(Zone zone);
   Future<List<Zone>> getZones();
-  Future<String> createProgram({
-    required String name,
-    required List<int> frequency,
-  });
-  Future<Program> getProgram({
-    required String programId,
-  });
+
+  // Programs
+  Future<void> insertProgram(Program program);
+  Future<void> updateProgram(Program program);
+  Future<Program> getProgram({required String programId});
   Future<List<Program>> getPrograms();
-  Future<void> updateProgram({
-    required String programId,
-    required String name,
-    required List<int> frequency,
-  });
-  Future<void> deleteProgram({
-    required String programId,
-  });
-  Future<void> insertRuns({
-    required String programId,
-    required List<Run> runs,
-  });
-  Future<void> updateRuns({
-    required String programId,
-    required List<Run> runs,
-  });
-  Future<List<Run>> getRunsForProgram({
-    required String programId,
-  });
-  Future<void> deleteRun({
-    required String runId,
-    required String programId,
-  });
+  Future<void> deleteProgram({required String programId});
+  Future<List<Run>> getRunsForProgram({required String programId});
+
+  // Misc.
   Future<void> clearAllData();
 }
