@@ -1,5 +1,6 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'developer_cubit.dart';
@@ -11,21 +12,7 @@ part of 'developer_cubit.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-/// @nodoc
-class _$DeveloperStateTearOff {
-  const _$DeveloperStateTearOff();
-
-  _DeveloperState call({required Map<String, dynamic> storage}) {
-    return _DeveloperState(
-      storage: storage,
-    );
-  }
-}
-
-/// @nodoc
-const $DeveloperState = _$DeveloperStateTearOff();
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 /// @nodoc
 mixin _$DeveloperState {
@@ -67,33 +54,33 @@ class _$DeveloperStateCopyWithImpl<$Res>
 }
 
 /// @nodoc
-abstract class _$DeveloperStateCopyWith<$Res>
+abstract class _$$_DeveloperStateCopyWith<$Res>
     implements $DeveloperStateCopyWith<$Res> {
-  factory _$DeveloperStateCopyWith(
-          _DeveloperState value, $Res Function(_DeveloperState) then) =
-      __$DeveloperStateCopyWithImpl<$Res>;
+  factory _$$_DeveloperStateCopyWith(
+          _$_DeveloperState value, $Res Function(_$_DeveloperState) then) =
+      __$$_DeveloperStateCopyWithImpl<$Res>;
   @override
   $Res call({Map<String, dynamic> storage});
 }
 
 /// @nodoc
-class __$DeveloperStateCopyWithImpl<$Res>
+class __$$_DeveloperStateCopyWithImpl<$Res>
     extends _$DeveloperStateCopyWithImpl<$Res>
-    implements _$DeveloperStateCopyWith<$Res> {
-  __$DeveloperStateCopyWithImpl(
-      _DeveloperState _value, $Res Function(_DeveloperState) _then)
-      : super(_value, (v) => _then(v as _DeveloperState));
+    implements _$$_DeveloperStateCopyWith<$Res> {
+  __$$_DeveloperStateCopyWithImpl(
+      _$_DeveloperState _value, $Res Function(_$_DeveloperState) _then)
+      : super(_value, (v) => _then(v as _$_DeveloperState));
 
   @override
-  _DeveloperState get _value => super._value as _DeveloperState;
+  _$_DeveloperState get _value => super._value as _$_DeveloperState;
 
   @override
   $Res call({
     Object? storage = freezed,
   }) {
-    return _then(_DeveloperState(
+    return _then(_$_DeveloperState(
       storage: storage == freezed
-          ? _value.storage
+          ? _value._storage
           : storage // ignore: cast_nullable_to_non_nullable
               as Map<String, dynamic>,
     ));
@@ -103,10 +90,15 @@ class __$DeveloperStateCopyWithImpl<$Res>
 /// @nodoc
 
 class _$_DeveloperState implements _DeveloperState {
-  _$_DeveloperState({required this.storage});
+  _$_DeveloperState({required final Map<String, dynamic> storage})
+      : _storage = storage;
 
+  final Map<String, dynamic> _storage;
   @override
-  final Map<String, dynamic> storage;
+  Map<String, dynamic> get storage {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableMapView(_storage);
+  }
 
   @override
   String toString() {
@@ -117,28 +109,28 @@ class _$_DeveloperState implements _DeveloperState {
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _DeveloperState &&
-            const DeepCollectionEquality().equals(other.storage, storage));
+            other is _$_DeveloperState &&
+            const DeepCollectionEquality().equals(other._storage, _storage));
   }
 
   @override
   int get hashCode =>
-      Object.hash(runtimeType, const DeepCollectionEquality().hash(storage));
+      Object.hash(runtimeType, const DeepCollectionEquality().hash(_storage));
 
   @JsonKey(ignore: true)
   @override
-  _$DeveloperStateCopyWith<_DeveloperState> get copyWith =>
-      __$DeveloperStateCopyWithImpl<_DeveloperState>(this, _$identity);
+  _$$_DeveloperStateCopyWith<_$_DeveloperState> get copyWith =>
+      __$$_DeveloperStateCopyWithImpl<_$_DeveloperState>(this, _$identity);
 }
 
 abstract class _DeveloperState implements DeveloperState {
-  factory _DeveloperState({required Map<String, dynamic> storage}) =
+  factory _DeveloperState({required final Map<String, dynamic> storage}) =
       _$_DeveloperState;
 
   @override
   Map<String, dynamic> get storage;
   @override
   @JsonKey(ignore: true)
-  _$DeveloperStateCopyWith<_DeveloperState> get copyWith =>
+  _$$_DeveloperStateCopyWith<_$_DeveloperState> get copyWith =>
       throw _privateConstructorUsedError;
 }
