@@ -1,5 +1,6 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'programs_cubit.dart';
@@ -11,24 +12,7 @@ part of 'programs_cubit.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-/// @nodoc
-class _$ProgramsStateTearOff {
-  const _$ProgramsStateTearOff();
-
-  _ProgramsState call(
-      {required List<Program> programs,
-      List<Program> pendingDeletes = const []}) {
-    return _ProgramsState(
-      programs: programs,
-      pendingDeletes: pendingDeletes,
-    );
-  }
-}
-
-/// @nodoc
-const $ProgramsState = _$ProgramsStateTearOff();
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 /// @nodoc
 mixin _$ProgramsState {
@@ -76,38 +60,38 @@ class _$ProgramsStateCopyWithImpl<$Res>
 }
 
 /// @nodoc
-abstract class _$ProgramsStateCopyWith<$Res>
+abstract class _$$_ProgramsStateCopyWith<$Res>
     implements $ProgramsStateCopyWith<$Res> {
-  factory _$ProgramsStateCopyWith(
-          _ProgramsState value, $Res Function(_ProgramsState) then) =
-      __$ProgramsStateCopyWithImpl<$Res>;
+  factory _$$_ProgramsStateCopyWith(
+          _$_ProgramsState value, $Res Function(_$_ProgramsState) then) =
+      __$$_ProgramsStateCopyWithImpl<$Res>;
   @override
   $Res call({List<Program> programs, List<Program> pendingDeletes});
 }
 
 /// @nodoc
-class __$ProgramsStateCopyWithImpl<$Res>
+class __$$_ProgramsStateCopyWithImpl<$Res>
     extends _$ProgramsStateCopyWithImpl<$Res>
-    implements _$ProgramsStateCopyWith<$Res> {
-  __$ProgramsStateCopyWithImpl(
-      _ProgramsState _value, $Res Function(_ProgramsState) _then)
-      : super(_value, (v) => _then(v as _ProgramsState));
+    implements _$$_ProgramsStateCopyWith<$Res> {
+  __$$_ProgramsStateCopyWithImpl(
+      _$_ProgramsState _value, $Res Function(_$_ProgramsState) _then)
+      : super(_value, (v) => _then(v as _$_ProgramsState));
 
   @override
-  _ProgramsState get _value => super._value as _ProgramsState;
+  _$_ProgramsState get _value => super._value as _$_ProgramsState;
 
   @override
   $Res call({
     Object? programs = freezed,
     Object? pendingDeletes = freezed,
   }) {
-    return _then(_ProgramsState(
+    return _then(_$_ProgramsState(
       programs: programs == freezed
-          ? _value.programs
+          ? _value._programs
           : programs // ignore: cast_nullable_to_non_nullable
               as List<Program>,
       pendingDeletes: pendingDeletes == freezed
-          ? _value.pendingDeletes
+          ? _value._pendingDeletes
           : pendingDeletes // ignore: cast_nullable_to_non_nullable
               as List<Program>,
     ));
@@ -117,13 +101,26 @@ class __$ProgramsStateCopyWithImpl<$Res>
 /// @nodoc
 
 class _$_ProgramsState implements _ProgramsState {
-  _$_ProgramsState({required this.programs, this.pendingDeletes = const []});
+  _$_ProgramsState(
+      {required final List<Program> programs,
+      final List<Program> pendingDeletes = const []})
+      : _programs = programs,
+        _pendingDeletes = pendingDeletes;
 
+  final List<Program> _programs;
   @override
-  final List<Program> programs;
+  List<Program> get programs {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_programs);
+  }
+
+  final List<Program> _pendingDeletes;
+  @override
   @JsonKey()
-  @override
-  final List<Program> pendingDeletes;
+  List<Program> get pendingDeletes {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_pendingDeletes);
+  }
 
   @override
   String toString() {
@@ -134,28 +131,28 @@ class _$_ProgramsState implements _ProgramsState {
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _ProgramsState &&
-            const DeepCollectionEquality().equals(other.programs, programs) &&
+            other is _$_ProgramsState &&
+            const DeepCollectionEquality().equals(other._programs, _programs) &&
             const DeepCollectionEquality()
-                .equals(other.pendingDeletes, pendingDeletes));
+                .equals(other._pendingDeletes, _pendingDeletes));
   }
 
   @override
   int get hashCode => Object.hash(
       runtimeType,
-      const DeepCollectionEquality().hash(programs),
-      const DeepCollectionEquality().hash(pendingDeletes));
+      const DeepCollectionEquality().hash(_programs),
+      const DeepCollectionEquality().hash(_pendingDeletes));
 
   @JsonKey(ignore: true)
   @override
-  _$ProgramsStateCopyWith<_ProgramsState> get copyWith =>
-      __$ProgramsStateCopyWithImpl<_ProgramsState>(this, _$identity);
+  _$$_ProgramsStateCopyWith<_$_ProgramsState> get copyWith =>
+      __$$_ProgramsStateCopyWithImpl<_$_ProgramsState>(this, _$identity);
 }
 
 abstract class _ProgramsState implements ProgramsState {
   factory _ProgramsState(
-      {required List<Program> programs,
-      List<Program> pendingDeletes}) = _$_ProgramsState;
+      {required final List<Program> programs,
+      final List<Program> pendingDeletes}) = _$_ProgramsState;
 
   @override
   List<Program> get programs;
@@ -163,6 +160,6 @@ abstract class _ProgramsState implements ProgramsState {
   List<Program> get pendingDeletes;
   @override
   @JsonKey(ignore: true)
-  _$ProgramsStateCopyWith<_ProgramsState> get copyWith =>
+  _$$_ProgramsStateCopyWith<_$_ProgramsState> get copyWith =>
       throw _privateConstructorUsedError;
 }
