@@ -1,6 +1,9 @@
+import 'package:api_models/api_models.dart';
+import 'package:core/core.dart';
+import 'package:irri/auth/auth.dart';
 import 'package:irri/customer_details/customer_details.dart';
 import 'package:irri/programs/programs.dart';
-import 'package:uuid/uuid.dart';
+import 'package:result_type/result_type.dart';
 
 /// Creates a [Program] in the provided [CustomerDetailsRepository]
 /// with the given name, frequency, and runs.
@@ -13,37 +16,55 @@ import 'package:uuid/uuid.dart';
 /// [RunGroup].
 class CreateProgram {
   CreateProgram({
+    required HttpClient httpClient,
+    required GetApiKey getApiKey,
     required CustomerDetailsRepository repository,
-  }) : _repository = repository;
+  })  : _httpClient = httpClient,
+        _getApiKey = getApiKey,
+        _repository = repository;
 
+  final HttpClient _httpClient;
+  final GetApiKey _getApiKey;
   final CustomerDetailsRepository _repository;
 
-  Future<void> call({
+  Future<UseCaseResult<String, String>> call({
     required String name,
     required List<int> frequency,
     required List<RunGroup> runGroups,
   }) async {
-    final programId = await _repository.createProgram(
-      name: name,
-      frequency: frequency,
-    );
-    final runs = <Run>[];
+    final apiKey = await _getApiKey();
+
+    final runs = <RunCreation>[];
     for (final group in runGroups) {
       for (final zoneId in group.zoneIds) {
         runs.add(
-          Run(
-            id: const Uuid().v4(),
-            programId: programId,
-            startTime: group.timeOfDay,
-            duration: group.duration.inSeconds,
+          RunCreation(
+            startHour: group.timeOfDay.hour,
+            startMinute: group.timeOfDay.minute,
+            durationSeconds: group.duration.inSeconds,
             zoneId: zoneId,
           ),
         );
       }
     }
-    await _repository.insertRuns(
-      programId: programId,
-      runs: runs,
+
+    final response = await _httpClient.post<Map<String, dynamic>>(
+      'program',
+      data: CreateProgramRequest(
+        apiKey: apiKey!,
+        programName: name,
+        frequency: frequency,
+        runs: runs,
+      ),
     );
+
+    if (response.isSuccess) {
+      final createProgramResponse =
+          CreateProgramResponse.fromJson(response.success!);
+      await _repository.insertProgram(createProgramResponse.program);
+      return Success('Created program');
+    }
+
+    return Failure(response.failure.message);
   }
 }
