@@ -1,6 +1,6 @@
+import 'package:api_models/api_models.dart';
 import 'package:bloc/bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:hydrawise/hydrawise.dart';
 import 'package:irri/run_zone/run_zone.dart';
 
 part 'run_zone_state.dart';
