@@ -6,18 +6,15 @@ void main() {
   group('IsLoggedIn', () {
     final storage = InMemoryStorage();
     final setApiKey = SetApiKey(storage);
-    final setFirebaseUid = SetFirebaseUid(storage);
     final subject = IsLoggedIn(
       getApiKey: GetApiKey(storage),
-      getFirebaseUid: GetFirebaseUid(storage),
     );
 
     setUp(storage.clearAll);
 
-    group('when api key and firebase uid are stored', () {
+    group('when api key is stored', () {
       setUp(() async {
         await setApiKey('fake-api-key');
-        await setFirebaseUid('fake-firebase-uid');
       });
       test('is returns true', () async {
         final isLoggedIn = await subject.call();
@@ -25,27 +22,7 @@ void main() {
       });
     });
 
-    group('when only api key is stored', () {
-      setUp(() async {
-        await setApiKey('fake-api-key');
-      });
-      test('is returns false', () async {
-        final isLoggedIn = await subject.call();
-        expect(isLoggedIn, isFalse);
-      });
-    });
-
-    group('when only firebase uid is stored', () {
-      setUp(() async {
-        await setFirebaseUid('fake-firebase-uid');
-      });
-      test('is returns false', () async {
-        final isLoggedIn = await subject.call();
-        expect(isLoggedIn, isFalse);
-      });
-    });
-
-    group('when neither api key or firebase uid are stored', () {
+    group('when api key is not stored', () {
       test('is returns false', () async {
         final isLoggedIn = await subject.call();
         expect(isLoggedIn, isFalse);
