@@ -1,6 +1,6 @@
+import 'package:api_models/api_models.dart';
 import 'package:core/core.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:hydrawise/hydrawise.dart';
 import 'package:irri/auth/auth.dart';
 import 'package:irri/customer_details/customer_details.dart';
 
@@ -9,12 +9,8 @@ void main() {
     final storage = InMemoryStorage();
     final repository = InMemoryCustomerDetailsRepository();
     final setApiKey = SetApiKey(storage);
-    final setFirebaseUid = SetFirebaseUid(storage);
     final logOut = LogOut(
       setApiKey: setApiKey,
-      unauthenticateWithFirebase: FakeUnauthenticateWithFirebase(
-        setFirebaseUid: setFirebaseUid,
-      ),
       customerDetailsRepository: repository,
     );
     test('it sets api_key to an empty string', () async {
@@ -27,21 +23,26 @@ void main() {
     test('it clears all data in the repository', () async {
       final zone = Zone(
         id: 1,
-        physicalNumber: 1,
+        number: 1,
         name: 'Fake Zone',
-        nextTimeOfWaterFriendly: '7:00',
-        secondsUntilNextRun: 60,
-        lengthOfNextRunTimeOrTimeRemaining: 500,
+        timeUntilNextRunSec: 60,
+        runLengthSec: 500,
       );
       await repository.insertZone(zone);
       final customer = Customer(
         activeControllerId: 1,
         customerId: 1,
         apiKey: 'fake-api-key',
-        lastStatusUpdate: 60,
       );
       await repository.insertCustomer(customer);
-      await repository.createProgram(name: 'fake-program', frequency: [1]);
+      await repository.insertProgram(
+        Program(
+          id: '12345',
+          name: 'Fake Program',
+          frequency: [DateTime.monday],
+          runs: List.empty(),
+        ),
+      );
       await repository.addFcmToken('fake-token');
 
       expect(await repository.getZones(), isNotEmpty);
