@@ -1,3 +1,4 @@
+import 'package:api_models/api_models.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:irri/customer_details/customer_details.dart';
@@ -12,59 +13,66 @@ void main() {
       updateProgram = UpdateProgram(repository: repository);
     });
     test('it updates the program name', () async {
-      final programId = await repository.createProgram(
-        name: 'Original',
-        frequency: [],
+      await repository.insertProgram(
+        Program(
+          id: '12345',
+          name: 'Original',
+          frequency: [],
+          runs: List.empty(),
+        ),
       );
 
       await updateProgram(
-        programId: programId,
+        programId: '12345',
         name: 'New',
         frequency: [],
         runGroups: [],
       );
 
       expect(
-        await repository
-            .getProgram(programId: programId)
-            .then((value) => value.name),
+        await repository.getProgram(programId: '12345').then((value) => value.name),
         'New',
       );
     });
 
     test('it updates the program frequency', () async {
-      final programId = await repository.createProgram(
-        name: 'Original',
-        frequency: [],
+      await repository.insertProgram(
+        Program(
+          id: '12345',
+          name: 'Original',
+          frequency: [],
+          runs: List.empty(),
+        ),
       );
 
       await updateProgram(
-        programId: programId,
+        programId: '12345',
         name: 'New',
         frequency: [DateTime.monday],
         runGroups: [],
       );
 
       expect(
-        await repository
-            .getProgram(programId: programId)
-            .then((value) => value.frequency.contains(DateTime.monday)),
+        await repository.getProgram(programId: '12345').then((value) => value.frequency.contains(DateTime.monday)),
         isTrue,
       );
     });
 
     test('it adds a run when a zone is added', () async {
-      final programId = await repository.createProgram(
-        name: 'Original',
-        frequency: [],
+      await repository.insertProgram(
+        Program(
+          id: '12345',
+          name: 'Original',
+          frequency: [],
+          runs: List.empty(),
+        ),
       );
 
-      final runsBefore =
-          await repository.getRunsForProgram(programId: programId);
+      final runsBefore = await repository.getRunsForProgram(programId: '12345');
       expect(runsBefore.length, 0);
 
       await updateProgram(
-        programId: programId,
+        programId: '12345',
         name: 'New',
         frequency: [DateTime.monday],
         runGroups: [
@@ -77,44 +85,44 @@ void main() {
         ],
       );
 
-      final runsAfter =
-          await repository.getRunsForProgram(programId: programId);
+      final runsAfter = await repository.getRunsForProgram(programId: '12345');
 
       expect(runsAfter.length, 1);
       expect(runsAfter.single.zoneId, 1);
     });
 
     test('it removes a run when a zone is removed', () async {
-      final programId = await repository.createProgram(
-        name: 'Original',
-        frequency: [],
-      );
-      await repository.insertRuns(
-        programId: programId,
-        runs: [
-          Run(
-            id: 'id-1',
-            programId: programId,
-            startTime: const TimeOfDay(hour: 9, minute: 9),
-            duration: 55,
-            zoneId: 1,
-          ),
-          Run(
-            id: 'id-2',
-            programId: programId,
-            startTime: const TimeOfDay(hour: 9, minute: 9),
-            duration: 55,
-            zoneId: 2,
-          ),
-        ],
+      await repository.insertProgram(
+        Program(
+          id: '12345',
+          name: 'Original',
+          frequency: [],
+          runs: [
+            Run(
+              id: 'id-1',
+              programId: '12345',
+              startHour: 9,
+              startMinute: 9,
+              durationSeconds: 55,
+              zoneId: 1,
+            ),
+            Run(
+              id: 'id-2',
+              programId: '12345',
+              startHour: 9,
+              startMinute: 9,
+              durationSeconds: 55,
+              zoneId: 2,
+            ),
+          ],
+        ),
       );
 
-      final runsBefore =
-          await repository.getRunsForProgram(programId: programId);
+      final runsBefore = await repository.getRunsForProgram(programId: '12345');
       expect(runsBefore.length, 2);
 
       await updateProgram(
-        programId: programId,
+        programId: '12345',
         name: 'New',
         frequency: [],
         runGroups: [
@@ -127,38 +135,37 @@ void main() {
         ],
       );
 
-      final runsAfter =
-          await repository.getRunsForProgram(programId: programId);
+      final runsAfter = await repository.getRunsForProgram(programId: '12345');
 
       expect(runsAfter.length, 1);
       expect(runsAfter.single.zoneId, 1);
     });
 
     test('it updates start time and duration of existing zones', () async {
-      final programId = await repository.createProgram(
-        name: 'Original',
-        frequency: [],
-      );
-      await repository.insertRuns(
-        programId: programId,
-        runs: [
-          Run(
-            id: 'id-1',
-            programId: programId,
-            startTime: const TimeOfDay(hour: 9, minute: 9),
-            duration: 55,
-            zoneId: 1,
-          ),
-        ],
+      await repository.insertProgram(
+        Program(
+          id: '12345',
+          name: 'Original',
+          frequency: [],
+          runs: [
+            Run(
+              id: 'id-1',
+              programId: '12345',
+              startHour: 9,
+              startMinute: 9,
+              durationSeconds: 55,
+              zoneId: 1,
+            ),
+          ],
+        ),
       );
 
-      final runsBefore =
-          await repository.getRunsForProgram(programId: programId);
+      final runsBefore = await repository.getRunsForProgram(programId: '12345');
       expect(runsBefore.single.startTime, const TimeOfDay(hour: 9, minute: 9));
-      expect(runsBefore.single.duration, 55);
+      expect(runsBefore.single.durationSeconds, 55);
 
       await updateProgram(
-        programId: programId,
+        programId: '12345',
         name: 'New',
         frequency: [],
         runGroups: [
@@ -171,11 +178,10 @@ void main() {
         ],
       );
 
-      final runsAfter =
-          await repository.getRunsForProgram(programId: programId);
+      final runsAfter = await repository.getRunsForProgram(programId: '12345');
 
       expect(runsAfter.single.startTime, const TimeOfDay(hour: 5, minute: 5));
-      expect(runsAfter.single.duration, 25);
+      expect(runsAfter.single.durationSeconds, 25);
     });
   });
 }
